refactor(agent-tracing): rename test files to PascalCase and clean up code
- Rename test files to match PascalCase convention
- Remove unused module.yml
- Clean up TracingService and ClaudeAgentTracer code
- Remove unnecessary dependencies from package.json
